+++ Runtime/ArrayExtensions.cs
@@ -1,0 +1,10 @@
+﻿namespace System
+{
+    public static class ArrayExtensions
+    {
+        public static bool IsWithinBounds(this Array array, int i)
+        {
+            return i >= 0 && i < array.Length;
+        }
+    }
+}