+++ Runtime/ListExtensions.cs
@@ -1,0 +1,21 @@
+﻿namespace System.Collections
+{
+    public static class ListExtensions
+    {
+        public static bool IsNullOrEmpty(this IList list)
+        {
+            return list == null || list.Count == 0;
+        }
+    }
+}
+
+namespace System.Collections.Generic
+{
+    public static class GenericListExtensions
+    {
+        public static bool IsNullOrEmpty<T>(this IList<T> list)
+        {
+            return list == null || list.Count == 0;
+        }
+    }
+}