+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [2.0.0] - 2020-03-05
+### Added
+- Array extensions
+    - IsWithinBounds
+- GameObject extensions
+    - SetParent
+- Linq extensions
+    - None
+- List extensions
+    - IsNullOrEmpty
+- Math extensions
+    - Clamp
+- Object extensions
+    - IsNull
+    - IsNotNull
+    - GetAssetPath
+    - GetAssetGuid
+
 ## [1.0.0] - 2020-02-25
 ### Added
 - String extensions